Split PartialCallGraph into constructor class and data structure
