+++ analyzer/javacg-opal/src/test/java/eu/fasten/analyzer/javacgopal/data/JSONUtilsTest.java
@@ -42,17 +42,17 @@ class JSONUtilsTest {
 
         var coordinate =
             new MavenCoordinate("com.github.shoothzj", "java-tool", "3.0.30.RELEASE", "jar");
-        graph = PartialCallGraph.createExtendedRevisionJavaCallGraph(coordinate,
+        graph = PartialCallGraphConstructor.createExtendedRevisionJavaCallGraph(coordinate,
             CGAlgorithm.CHA, 1574072773, MavenUtilities.MAVEN_CENTRAL_REPO, CallPreservationStrategy.ONLY_STATIC_CALLSITES);
 
         coordinate =
             new MavenCoordinate("abbot", "costello", "1.4.0", "jar");
-        artifact = PartialCallGraph.createExtendedRevisionJavaCallGraph(coordinate,
+        artifact = PartialCallGraphConstructor.createExtendedRevisionJavaCallGraph(coordinate,
             CGAlgorithm.CHA, 1574072773, MavenUtilities.MAVEN_CENTRAL_REPO, CallPreservationStrategy.ONLY_STATIC_CALLSITES);
 
         coordinate =
             new MavenCoordinate("abbot", "abbot", "1.4.0", "jar");
-        dependency = PartialCallGraph.createExtendedRevisionJavaCallGraph(coordinate,
+        dependency = PartialCallGraphConstructor.createExtendedRevisionJavaCallGraph(coordinate,
             CGAlgorithm.CHA, 1574072773, MavenUtilities.MAVEN_CENTRAL_REPO, CallPreservationStrategy.ONLY_STATIC_CALLSITES);
         final var deps = new ArrayList<>(Collections.singletonList(dependency));
         deps.add(artifact);
@@ -101,7 +101,7 @@ class JSONUtilsTest {
 
         for (int i = 0; i < coordsSize; i++) {
             MavenCoordinate coord = coords.get(i);
-            final var cg = PartialCallGraph.createExtendedRevisionJavaCallGraph(coord,
+            final var cg = PartialCallGraphConstructor.createExtendedRevisionJavaCallGraph(coord,
                 CGAlgorithm.CHA, 1574072773, MavenUtilities.getRepos().get(0), CallPreservationStrategy.ONLY_STATIC_CALLSITES);
 
             logger.debug("Serialization for: {}", coord.getCoordinate());
